+++ src/main/java/raisetech/SutudentManagement/controller/StudentController.java
@@ -1,0 +1,30 @@
+package raisetech.SutudentManagement.controller;
+
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+import raisetech.SutudentManagement.data.Student;
+import raisetech.SutudentManagement.data.StudentCourse;
+import raisetech.SutudentManagement.service.StudentService;
+
+@RestController
+public class StudentController {
+
+  private StudentService service;
+
+  @Autowired
+  public StudentController(StudentService service) {
+    this.service = service;
+  }
+
+  @GetMapping("/studentList")
+  public List<Student> getStudentList() {
+    return service.searchStudentList();
+  }
+
+  @GetMapping("/studentCourseList")
+  public List<StudentCourse> getStudentCourseList() {
+    return service.searchStudentCourseList();
+  }
+}
