+++ src/main/java/raisetech/SutudentManagement/repository/StudentRepository.java
@@ -1,8 +1,10 @@
-package raisetech.SutudentManagement;
+package raisetech.SutudentManagement.repository;
 
 import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
+import raisetech.SutudentManagement.data.Student;
+import raisetech.SutudentManagement.data.StudentCourse;
 
 @Mapper
 public interface StudentRepository {
@@ -11,5 +13,5 @@ public interface StudentRepository {
   List<Student> search();
 
   @Select("SELECT * FROM students_courses")
-  List<StudentCourse> find();
+  List<StudentCourse> searchStudentsCourses();
 }
