+++ src/main/java/raisetech/SutudentManagement/service/StudentService.java
@@ -1,0 +1,28 @@
+package raisetech.SutudentManagement.service;
+
+import java.util.List;
+import java.util.stream.Collectors;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import raisetech.SutudentManagement.data.Student;
+import raisetech.SutudentManagement.data.StudentCourse;
+import raisetech.SutudentManagement.repository.StudentRepository;
+
+@Service
+public class StudentService {
+
+  private StudentRepository repository;
+
+  @Autowired
+  public StudentService(StudentRepository repository) {
+    this.repository = repository;
+  }
+
+  public List<Student> searchStudentList() {
+    return repository.search();
+  }
+
+  public List<StudentCourse> searchStudentCourseList() {
+    return repository.searchStudentsCourses();
+  }
+}
